// apps/web/src/lib/api/model/purchaseOrdersEvaluationsIndexStatus.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Procurement
 * OpenAPI spec version: 0.0.1
 */

export type PurchaseOrdersEvaluationsIndexStatus = typeof PurchaseOrdersEvaluationsIndexStatus[keyof typeof PurchaseOrdersEvaluationsIndexStatus] | null;


export const PurchaseOrdersEvaluationsIndexStatus = {
  draft: 'draft',
  in_review: 'in_review',
  approved: 'approved',
  rejected: 'rejected',
  returned: 'returned',
} as const;